requirements.txt
altgraph==0.17
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
docopt==0.6.2
future==0.18.2
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.4.0
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.2
parsel==1.6.0
pefile==2019.4.18
Pillow==8.1.0
pipreqs==0.4.10
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyOpenSSL==20.0.1
pywin32-ctypes==0.2.0
queuelib==1.5.0
requests==2.25.1
resolver==0.2.1
Scrapy==2.4.1
selector==0.10.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
tencentcloud-sdk-python==3.0.325
Twisted==20.3.0
typing-extensions==3.7.4.3
urllib3==1.26.2
w3lib==1.22.0
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
zipp==3.4.0
zope.interface==5.2.0
